corrected field

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -6,7 +6,7 @@
 */
 
 
-int _strlen(char *s)
+size_t _strlen(char *s)
 {
 int i = 0;
 while (s[i])
